+++ bin/store/CHANGELOG.md
@@ -12,6 +12,9 @@ Bert
 ### Fixes
 - ***validations.rs*** revise validation for distinct_by that causes the app to crash
 - Main table selections are not shown if there's a join
+- set `aliased_entity` to optional in `concatenate_fields`
+- fix concatenated field names validation for main fields
+- fix the concatenated fields for main fields
 ---
 
 ## 0.1.21